delete localised to trash icon only

+++ public/javascript/client.js
@@ -10,18 +10,24 @@ function handleClick(e) {
     const todoTag = e.target.closest('.todo-item');
     const todoItem = todoTag.getElementsByClassName('todo')[0].innerText;
 
-    const clickeTodo = {
-        todoItem: todoItem
-    };
-
-    fetch('http://localhost:3000/', {
-        method: 'DELETE',
-        body: JSON.stringify(clickeTodo),
-        headers: { 'Content-type': 'application/json; charset=UTF-8' }
-    })
-        .then(response => {
-            window.location = response.url;
-            return response.json();
+    if (e.target.className === 'trash') {
+        const clickeTodo = {
+            todoItem: todoItem
+        };
+        fetch('http://localhost:3000/', {
+            method: 'DELETE',
+            body: JSON.stringify(clickeTodo),
+            headers: { 'Content-type': 'application/json; charset=UTF-8' }
         })
-        .catch(err => console.log(err));
+            .then(response => {
+                window.location = response.url;
+                return response.json();
+            })
+            .catch(err => console.log(err));
+    }
+
+
+
+
+
 }
